Added a helper to display facets as checkboxes.
