+++ config/module.config.php
@@ -30,6 +30,7 @@ return [
             'apiSearch' => Service\ViewHelper\ApiSearchFactory::class,
             'apiSearchOne' => Service\ViewHelper\ApiSearchOneFactory::class,
             'facetLabel' => Service\ViewHelper\FacetLabelFactory::class,
+            'facetCheckbox' => Service\ViewHelper\FacetCheckboxFactory::class,
             'facetLink' => Service\ViewHelper\FacetLinkFactory::class,
             'searchIndexConfirm' => Service\ViewHelper\SearchIndexConfirmFactory::class,
         ],
