+++ src/View/Helper/FacetCheckbox.php
@@ -1,0 +1,7 @@
+<?php
+namespace Search\View\Helper;
+
+class FacetCheckbox extends AbstractFacetElement
+{
+    protected $partial = 'search/facet-checkbox';
+}
