+++ view/search/facet-checkbox.phtml
@@ -1,0 +1,17 @@
+<?php
+/**
+ * @var \Zend\View\Renderer\PhpRenderer $this
+ * @var string $name
+ * @var string $value
+ * @var string $url
+ * @var int $count
+ * @var bool $active
+ * @var \Zend\View\Helper\EscapeHtml $escapeHtml
+ */
+
+$escapedValue = $escapeHtml($value);
+$escapedId = $name . '[]' . $escapedValue;
+?>
+
+<input type="checkbox" name="<?= $name ?>[]" id="<?= $escapedId ?>" value="<?= $escapedValue ?>" data-url="<?= $url ?>" <?php if ($active) echo ' checked="checked"'; ?>/>
+<label for="<?= $escapedId ?>"><?= $escapedValue ?> (<span class="count"><?= $count ?></span>)</label>
